% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRoastProfiles.R
\name{compareRoastProfiles}
\alias{compareRoastProfiles}
\title{compareRoastProfiles}
\usage{
compareRoastProfiles(file_list, degree = 8, names = NULL)
}
\arguments{
\item{file_list}{A vector of roast files.}

\item{degree}{The degree of the polynomial used for smoothing temp vs. time}

\item{names}{The names of the roasts (for the legend).}
}
\description{
A function that will compare roast profiles for a number of CSV files.
}
