% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRoastProfile.R
\name{plotRoastProfile}
\alias{plotRoastProfile}
\title{plotRoastProfile}
\usage{
plotRoastProfile(filename, labels = TRUE, delta = TRUE, degree = 8,
  combined = FALSE)
}
\arguments{
\item{filename}{The filename of the roast profile}

\item{labels}{Logical - plot labels or not}

\item{delta}{Logical - plot delta bean temp or not.}
}
\description{
Plot a single roast profile - takes the filename of a filled out profile
template (found in /data). It takes the total_seconds and temp column to
plot the profile, the target column to plot the targets, and uses the notes
column to plot points in the roast where any note was taken. Uses the note
as a label, unless the notes are standard - e.g. fcs becomes a label
start of first crack. Rate of rise max will be calculated.
Also plots the delta bean temp, which is, I think, the first differential
of the curve?
}
